Don't set killing spree player if alerts are off

+++ source/blood/src/player.cpp
@@ -2197,7 +2197,7 @@ void playerFrag(PLAYER *pKiller, PLAYER *pVictim)
             if (bKillerSpreeActive && bKillerAlive) // if killed enemy within multi kill time window, reward point
             {
                 gMultiKillsFrags[nKiller]++;
-                if ((pKiller != gMe) && ((gMultiKillsFrags[nKiller] % 5) == 0)) // announce killing spree every 5 kills
+                if (gMultiKill && (pKiller != gMe) && ((gMultiKillsFrags[nKiller] % 5) == 0)) // announce killing spree every 5 kills
                 {
                     if ((gMultiKill == 2) && (gAnnounceKillingSpreeTicks == 0)) // only play sfx if multi kill alert setting is on, and if there isn't a multi kill alert already active
                         sndStartSample("NOTBLOOD6", 128, -1, 22050);
